+++ pycep/formatter.py
@@ -26,10 +26,13 @@ def add_whitespace(input_string):
 def strip_unsafe_file_names(string_data: str) -> str:
     new_string = ""
     for character in string_data:
-        if character != "/":
+        if character != "/" and character != ":":
             new_string += character
+        elif character == ":":
+            new_string += ";"
         else:
             new_string += "-"
+
     return new_string